requirements.txt
certifi==2019.6.16
chardet==3.0.4
idna==2.8
numpy==1.17.1
paho-mqtt==1.4.0
requests==2.22.0
urllib3==1.25.3
